+++ integration-tests/src/test/java/tech/beshu/ror/integration/IndicesReverseWildcardTests.java
@@ -14,7 +14,7 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.integration.other;
+package tech.beshu.ror.integration;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpGet;
