+++ integration-tests/src/test/java/tech/beshu/ror/integration/LdapIntegrationFirstOptionTests.java
@@ -14,13 +14,13 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.integration.other;
+package tech.beshu.ror.integration;
 
 import tech.beshu.ror.utils.containers.ESWithReadonlyRestContainer;
 import tech.beshu.ror.utils.containers.ESWithReadonlyRestContainerUtils;
+import tech.beshu.ror.utils.containers.LdapContainer;
 import tech.beshu.ror.utils.containers.MultiContainer;
 import tech.beshu.ror.utils.containers.MultiContainerDependent;
-import tech.beshu.ror.utils.containers.WireMockContainer;
 import tech.beshu.ror.utils.gradle.RorPluginGradleProject;
 import tech.beshu.ror.utils.integration.ElasticsearchTweetsInitializer;
 import tech.beshu.ror.utils.integration.ReadonlyRestedESAssertions;
@@ -29,36 +29,43 @@ import org.junit.Test;
 
 import static tech.beshu.ror.utils.integration.ReadonlyRestedESAssertions.assertions;
 
-public class ExternalAuthenticationTests {
+public class LdapIntegrationFirstOptionTests {
 
   @ClassRule
   public static MultiContainerDependent<ESWithReadonlyRestContainer> container =
     ESWithReadonlyRestContainerUtils.create(
       RorPluginGradleProject.fromSystemProperty(),
       new MultiContainer.Builder()
-        .add("EXT1", () -> WireMockContainer.create("/external_authentication/wiremock_service1_cartman.json",
-          "/external_authentication/wiremock_service1_morgan.json"
-        ))
-        .add("EXT2", () -> WireMockContainer.create("/external_authentication/wiremock_service2_cartman.json"))
-        .build(), "/external_authentication/elasticsearch.yml",
+        .add("LDAP1", () -> LdapContainer.create("/ldap_integration_1st/ldap.ldif"))
+        .add("LDAP2", () -> LdapContainer.create("/ldap_integration_1st/ldap.ldif"))
+        .build(),
+      "/ldap_integration_1st/elasticsearch.yml",
       new ElasticsearchTweetsInitializer()
     );
 
-
   @Test
-  public void testAuthenticationSuccessWithService1() throws Exception {
-    assertions(container).assertUserHasAccessToIndex("cartman", "user1", "twitter");
-  }
-
-  @Test
-  public void testAuthenticationErrorWithService1() throws Exception {
+  public void usersFromGroup1CanSeeTweets() throws Exception {
     ReadonlyRestedESAssertions assertions = assertions(container);
-    assertions.assertUserAccessToIndexForbidden("cartman", "user2", "twitter");
-    assertions.assertUserAccessToIndexForbidden("morgan", "user2", "twitter");
+    assertions.assertUserHasAccessToIndex("cartman", "user2", "twitter");
+    assertions.assertUserHasAccessToIndex("bong", "user1", "twitter");
   }
 
   @Test
-  public void testAuthenticationSuccessWithService2() throws Exception {
-    assertions(container).assertUserHasAccessToIndex("cartman", "user1", "facebook");
+  public void usersFromOutsideOfGroup1CannotSeeTweets() throws Exception {
+    assertions(container).assertUserAccessToIndexForbidden("morgan", "user1", "twitter");
   }
+
+  @Test
+  public void unauthenticatedUserCannotSeeTweets() throws Exception {
+    assertions(container).assertUserAccessToIndexForbidden("cartman", "wrong_password", "twitter");
+  }
+
+  @Test
+  public void usersFromGroup3CanSeeFacebookPosts() throws Exception {
+    ReadonlyRestedESAssertions assertions = assertions(container);
+    assertions.assertUserHasAccessToIndex("cartman", "user2", "facebook");
+    assertions.assertUserHasAccessToIndex("bong", "user1", "facebook");
+    assertions.assertUserHasAccessToIndex("morgan", "user1", "facebook");
+  }
+
 }
