+++ integration-tests/src/test/java/tech/beshu/ror/integration/LdapIntegrationWithLocalGroupsTests.java
@@ -15,7 +15,7 @@
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
 
-package tech.beshu.ror.integration.ldap;
+package tech.beshu.ror.integration;
 
 import com.google.common.base.Joiner;
 import org.apache.http.Header;
@@ -35,44 +35,27 @@ import tech.beshu.ror.utils.integration.ReadonlyRestedESAssertions;
 import static org.junit.Assert.assertEquals;
 import static tech.beshu.ror.utils.integration.ReadonlyRestedESAssertions.assertions;
 
-public class LdapIntegrationGroupsHeadersTests {
+public class LdapIntegrationWithLocalGroupsTests {
 
   @ClassRule
-  public static MultiContainerDependent<ESWithReadonlyRestContainer> container =
+  public static MultiContainerDependent<ESWithReadonlyRestContainer> container2 =
       ESWithReadonlyRestContainerUtils.create(
           RorPluginGradleProject.fromSystemProperty(),
           new MultiContainer.Builder()
-              .add("LDAP1", () -> LdapContainer.create("/ldap_integration_group_headers/ldap.ldif"))
+              .add("LDAP1", () -> LdapContainer.create("/ldap_separate_authc_authz_mixed_local/ldap.ldif"))
               .build(),
-          "/ldap_integration_group_headers/elasticsearch.yml",
+          "/ldap_separate_authc_authz_mixed_local/elasticsearch.yml",
           new ElasticsearchTweetsInitializer()
       );
 
-
-  @Test
-  public void checkCartmanRespHeadersWithoutCurrGroupHeader() throws Exception {
-    ReadonlyRestedESAssertions assertions = assertions(container);
-    assertions.assertUserHasAccessToIndex("cartman", "user2", "twitter", response -> {
-          System.out.println("resp headers" + Joiner.on(",").join(response.getAllHeaders()));
-          assertEquals("group1,group3,groupAll", getHeader(Constants.HEADER_GROUPS_AVAILABLE, response));
-          assertEquals("cartman", getHeader(Constants.HEADER_USER_ROR,response));
-          assertEquals("group1", getHeader(Constants.HEADER_GROUP_CURRENT,response));
-          assertEquals(".kibana_group1", getHeader(Constants.HEADER_KIBANA_INDEX, response));
-          return null;
-        },
-        httpRequest -> {
-          return null;
-        });
-  }
-
   @Test
   public void checkCartmanRespHeaders() throws Exception {
-    ReadonlyRestedESAssertions assertions = assertions(container);
+    ReadonlyRestedESAssertions assertions = assertions(container2);
     assertions.assertUserHasAccessToIndex("cartman", "user2", "twitter", response -> {
           System.out.println("resp headers" + Joiner.on(",").join(response.getAllHeaders()));
-          assertEquals("group1,group3,groupAll", getHeader(Constants.HEADER_GROUPS_AVAILABLE, response));
-          assertEquals("cartman", getHeader(Constants.HEADER_USER_ROR,response));
-          assertEquals("group1", getHeader(Constants.HEADER_GROUP_CURRENT,response));
+          assertEquals("group1,group3", getHeader(Constants.HEADER_GROUPS_AVAILABLE, response));
+          assertEquals("cartman", getHeader(Constants.HEADER_USER_ROR, response));
+          assertEquals("group1", getHeader(Constants.HEADER_GROUP_CURRENT, response));
           assertEquals(".kibana_group1", getHeader(Constants.HEADER_KIBANA_INDEX, response));
           return null;
         },
@@ -82,14 +65,15 @@ public class LdapIntegrationGroupsHeadersTests {
         });
   }
 
+
   @Test
   public void checkCartmanRespHeadersWithCurrentGroupReqHeader() throws Exception {
-    ReadonlyRestedESAssertions assertions = assertions(container);
+    ReadonlyRestedESAssertions assertions = assertions(container2);
     assertions.assertUserHasAccessToIndex("cartman", "user2", "twitter", response -> {
           System.out.println("resp headers" + Joiner.on(",\n").join(response.getAllHeaders()));
-          assertEquals("group1,group3,groupAll", getHeader(Constants.HEADER_GROUPS_AVAILABLE, response));
-          assertEquals("cartman", getHeader(Constants.HEADER_USER_ROR,response));
-          assertEquals("group3", getHeader(Constants.HEADER_GROUP_CURRENT,response));
+          assertEquals("group1,group3", getHeader(Constants.HEADER_GROUPS_AVAILABLE, response));
+          assertEquals("cartman", getHeader(Constants.HEADER_USER_ROR, response));
+          assertEquals("group3", getHeader(Constants.HEADER_GROUP_CURRENT, response));
           assertEquals(".kibana_group3", getHeader(Constants.HEADER_KIBANA_INDEX, response));
           return null;
         },
@@ -99,12 +83,11 @@ public class LdapIntegrationGroupsHeadersTests {
         });
   }
 
-  private String getHeader(String headerName, HttpResponse resp){
+  private String getHeader(String headerName, HttpResponse resp) {
     Header h = resp.getFirstHeader(headerName);
-    if (h == null){
+    if (h == null) {
       return "";
     }
     return h.getValue();
   }
-
 }
