+++ integration-tests/src/test/java/tech/beshu/ror/integration/RorKbnAuthTests.java
@@ -15,9 +15,8 @@
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
 
-package tech.beshu.ror.integration.other;
+package tech.beshu.ror.integration;
 
-import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
@@ -30,15 +29,13 @@ import tech.beshu.ror.utils.containers.ESWithReadonlyRestContainer;
 import tech.beshu.ror.utils.gradle.RorPluginGradleProject;
 import tech.beshu.ror.utils.httpclient.RestClient;
 
-import java.util.Date;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
 
-public class JwtAuthTests {
+public class RorKbnAuthTests {
 
   private static final String ALGO = "HS256";
   private static final String KEY = "123456.123456.123456.123456.123456.123456.123456.123456.123456.123456.123456.123456.123456.123456.123456.123456";
@@ -46,13 +43,13 @@ public class JwtAuthTests {
   private static final String WRONG_KEY = "abcdef";
   private static final String SUBJECT = "test";
   private static final String USER_CLAIM = "user";
-  private static final String ROLES_CLAIM = "roles";
+  private static final String GROUPS_CLAIM = "groups";
   private static final String EXP = "exp";
 
   @ClassRule
   public static ESWithReadonlyRestContainer container =
       ESWithReadonlyRestContainer.create(
-          RorPluginGradleProject.fromSystemProperty(), "/jwt_auth/elasticsearch.yml",
+          RorPluginGradleProject.fromSystemProperty(), "/ror_kbn_auth/elasticsearch.yml",
           Optional.empty()
       );
 
@@ -76,26 +73,9 @@ public class JwtAuthTests {
 
   @Test
   public void acceptValidTokentWithUserClaim() throws Exception {
-    int sc = test(makeToken(KEY, makeClaimMap(USER_CLAIM, "user")));
+    // Groups claim is mandatory, even if empty
+    int sc = test(makeToken(KEY, makeClaimMap(USER_CLAIM, "user", GROUPS_CLAIM, "")));
     assertEquals(200, sc);
-  }
-
-  @Test
-  public void acceptValidTokentWithUserClaimAndCustomHeader() throws Exception {
-    int sc = test(makeToken(KEY, makeClaimMap(USER_CLAIM, "user")), Optional.of("x-custom-header"), false);
-    assertEquals(200, sc);
-  }
-
-  @Test
-  public void acceptValidTokentWithUserClaimAndCustomHeaderAndCustomHeaderPrefix() throws Exception {
-    int sc = test(makeToken(KEY, makeClaimMap(USER_CLAIM, "user")), Optional.of("x-custom-header2"), Optional.of("x-custom-prefix"));
-    assertEquals(200, sc);
-  }
-
-  @Test
-  public void rejectTokentWithUserClaimAndCustomHeader() throws Exception {
-    int sc = test(makeToken(KEY, makeClaimMap("inexistent_claim", "user")), Optional.of("x-custom-header"), false);
-    assertEquals(401, sc);
   }
 
   @Test
@@ -112,17 +92,13 @@ public class JwtAuthTests {
 
   @Test
   public void rejectTokenWithWrongRolesClaim() throws Exception {
-    int sc = test(makeToken(KEY_ROLE, makeClaimMap(ROLES_CLAIM, "role_wrong")));
+    int sc = test(makeToken(KEY_ROLE, makeClaimMap(GROUPS_CLAIM, "wrong_group")));
     assertEquals(401, sc);
   }
 
   @Test
   public void acceptValidTokentWithRolesClaim() throws Exception {
-    List<String> roles = Lists.newArrayList();
-    roles.add("role_viewer");
-    roles.add("something_lol");
-    Optional<String> tok = makeToken(KEY_ROLE, makeClaimMap(USER_CLAIM, "user1", ROLES_CLAIM, roles));
-    int sc = test(tok);
+    int sc = test(makeToken(KEY_ROLE, makeClaimMap(GROUPS_CLAIM, "viewer_group")));
     assertEquals(200, sc);
   }
 
@@ -131,13 +107,11 @@ public class JwtAuthTests {
   }
 
   private int test(Optional<String> token, Optional<String> headerName, boolean withBearer) throws Exception {
-    return test(token, headerName, Optional.of(withBearer ? "Bearer " : ""));
-  }
-
-  private int test(Optional<String> token, Optional<String> headerName, Optional<String> headerPrefix) throws Exception {
     RestClient rc = container.getClient();
     HttpGet req = new HttpGet(rc.from("/_cat/indices"));
-    token.ifPresent(t -> req.addHeader(headerName.orElse("Authorization"), headerPrefix.orElse("") + t));
+    token.ifPresent(t -> req.addHeader(headerName.orElse("Authorization"), (withBearer ? "Bearer " : "") + t));
+    System.out.println("sending request with auth header: " + req.getFirstHeader("Authorization"));
+
     HttpResponse resp = rc.execute(req);
     return resp.getStatusLine().getStatusCode();
   }
@@ -149,9 +123,7 @@ public class JwtAuthTests {
   private Optional<String> makeToken(String key, Map<String, Object> claims) {
     JwtBuilder builder = Jwts.builder()
                              .setSubject(SUBJECT)
-                             .setExpiration(new Date(System.currentTimeMillis() * 2))
                              .signWith(SignatureAlgorithm.valueOf(ALGO), key.getBytes());
-
     claims.forEach(builder::claim);
     return Optional.of(builder.compact());
   }
@@ -159,8 +131,12 @@ public class JwtAuthTests {
   private Map<String, Object> makeClaimMap(Object... kvs) {
     assert kvs.length % 2 == 0;
     HashMap<String, Object> claims = Maps.newHashMap();
-    for (int i = 0; i < kvs.length; i += 2)
+    for (int i = 0; i < kvs.length; i += 2) {
       claims.put((String) kvs[i], kvs[i + 1]);
+    }
+    if(!claims.containsKey(USER_CLAIM)){
+      claims.put(USER_CLAIM, "user");
+    }
     return claims;
   }
 }
